Import version 2.09
This was released 2012-06-23.  It appears that this was considered
a beta release at the time.

While this package is available from an upstream bazaar repository,
that repository does not contain any meaningful history.  It just
contains imports of a few dozen releases and each release is split
across hundreds of commits, one per file.

Also see https://answers.launchpad.net/nxhtml/+question/236477.